Update path in files utils

+++ utils/files.go
@@ -6,9 +6,12 @@ import (
 	"os"
 	"strings"
 
+	"github.com/dropbox/godropbox/container/set"
 	"github.com/dropbox/godropbox/errors"
 	"github.com/pritunl/pritunl-cloud/errortypes"
 )
+
+var invalidPaths = set.NewSet("/", "", ".", "./")
 
 func Chmod(pth string, mode os.FileMode) (err error) {
 	err = os.Chmod(pth, mode)
@@ -115,6 +118,13 @@ func ExistsRemove(pth string) (err error) {
 }
 
 func Remove(path string) (err error) {
+	if invalidPaths.Contains(path) {
+		err = &errortypes.WriteError{
+			errors.Wrapf(err, "utils: Invalid remove path '%s'", path),
+		}
+		return
+	}
+
 	err = os.Remove(path)
 	if err != nil {
 		err = &errortypes.WriteError{
@@ -127,6 +137,13 @@ func Remove(path string) (err error) {
 }
 
 func RemoveAll(path string) (err error) {
+	if invalidPaths.Contains(path) {
+		err = &errortypes.WriteError{
+			errors.Wrapf(err, "utils: Invalid remove path '%s'", path),
+		}
+		return
+	}
+
 	err = os.RemoveAll(path)
 	if err != nil {
 		err = &errortypes.WriteError{
